chore(cli): keep Vitest APIs out of the live globalSetup
Split the env-only live helpers (isLiveConfigured, liveApiBaseUrl, project-ref
accessors, profile/timeout constants) into tests/helpers/live-env.ts, which
imports no Vitest test APIs. globalSetup now imports from there instead of
helpers/live.ts (which pulls in `describe`), avoiding evaluating Vitest test
APIs in the globalSetup context. live.ts re-exports the env helpers so
*.live.test.ts keep a single import site.

Validated against a real supabox stack: 8 passed, 0 skipped.
